% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_pca.R
\name{alignment_pca}
\alias{alignment_pca}
\title{Perform PCA on Sequence Alignment}
\usage{
alignment_pca(
  alignment,
  exclude_gaps = TRUE,
  min_freq = 0,
  center = TRUE,
  scale = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{alignment}{An alignment object with 'seq' and 'nam' components}

\item{exclude_gaps}{Logical. If TRUE, gaps ("-") are excluded from the binary encoding. Default is TRUE.}

\item{min_freq}{Numeric between 0 and 1. Minimum frequency for a variant to be included. Default is 0.}

\item{center}{Logical. If TRUE, the variables are centered before PCA. Default is TRUE.}

\item{scale}{Logical. If TRUE, the variables are scaled before PCA. Default is TRUE.}

\item{verbose}{Logical. If TRUE, progress messages are printed. Default is FALSE.}
}
\value{
A list with components:
  \describe{
    \item{pca}{The result of prcomp}
    \item{binary_matrix}{The binary matrix used for PCA}
    \item{var_explained}{The proportion of variance explained by each principal component}
  }
}
\description{
This function converts a sequence alignment to a binary matrix and performs
principal component analysis (PCA) on it.
}
\examples{
\dontrun{
# Read an alignment file
library(seqinr)
my_alignment <- read.alignment(file = "my_alignment.fasta", format = "fasta")

# Perform PCA
pca_result <- alignment_pca(my_alignment, min_freq = 0.01)

# Plot the first two principal components
plot(pca_result$pca$x[,1], pca_result$pca$x[,2],
     xlab = "PC1", ylab = "PC2",
     main = "PCA of sequence alignment")

# Plot variance explained
barplot(pca_result$var_explained[1:10],
        names.arg = paste0("PC", 1:10),
        main = "Scree Plot",
        ylab = "Proportion of Variance Explained")
}

}
