% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgwas.R
\name{analyzePosition}
\alias{analyzePosition}
\title{Helper function to analyze a single position}
\usage{
analyzePosition(
  temp.Var,
  temp.idx,
  temp.aa,
  aa_index,
  data,
  temp.df,
  formula,
  results
)
}
\arguments{
\item{temp.Var}{Current variable name}

\item{temp.idx}{Index of current variable in data}

\item{temp.aa}{Table of amino acid frequencies}

\item{aa_index}{Index of the amino acid being analyzed}

\item{data}{Original data}

\item{temp.df}{Working data frame}

\item{formula}{Formula for the model}

\item{results}{Results data frame to append to}
}
\value{
Updated results data frame
}
\description{
Helper function to analyze a single position
}
\keyword{internal}
