% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr-data.R
\docType{data}
\name{chr-data}
\alias{chr-data}
\alias{chr}
\title{Chromosome Indices for The Markers of The Simulated Data}
\format{A numeric vector of chromosome indices for the 20K
simulated markers.}
\usage{
data(chr)
}
\description{
Chromosome indices for the markers of the simulated data
}
\examples{
data("chr", package="vGWAS")
table(chr)
}
\keyword{datasets}
