\name{chr}
\alias{chr}
\docType{data}
\title{
Chromosome Indices for The Markers of The Simulated Data
}
\description{
Chromosome indices for the markers of the simulated data
}
\usage{data(chr)}
\format{
  A numeric vector of chromosome indices for the 20K simulated markers.
}
\examples{

data(chr)
table(chr)

}