% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno-data.R
\docType{data}
\name{geno-data}
\alias{geno-data}
\alias{geno}
\title{The Marker Genotypes of The Simulated Data}
\format{A character matrix of size (number of individuals)
times (number of markers in the genome).}
\usage{
data(geno)
}
\description{
The marker genotypes of the simulated data
}
\details{
Note that there is only one column for each marker.
}
\examples{
data("geno", package="vGWAS")
}
\keyword{datasets}
