% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.df-data.R
\docType{data}
\name{geno.df-data}
\alias{geno.df-data}
\alias{geno.df}
\title{The Marker Genotypes of The Simulated Data (as data.frame)}
\format{
A data.frame of size (number of markers in the genome)
times (number of individuals).
}
\usage{
data(geno.df)
}
\description{
The marker genotypes of the simulated data
}
\details{
Note that there is only one column for each marker.
}
\examples{
data("geno.df", package="vGWAS")
}
\keyword{datasets}
