% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.num-data.R
\docType{data}
\name{geno.num-data}
\alias{geno.num-data}
\alias{geno.num}
\title{The Marker Genotypes of The Simulated Data (as matrix)}
\format{
A matrix of size (number of individuals)
times (number of markers in the genome).
}
\usage{
data(geno.num)
}
\description{
The marker genotypes of the simulated data
}
\details{
Note that there is only one column for each marker.
}
\examples{
data("geno.num", package="vGWAS")
}
\keyword{datasets}
