\name{geno}
\alias{geno}
\docType{data}
\title{
The Marker Genotypes of The Simulated Data
}
\description{
The marker genotypes of the simulated data
}
\usage{data(geno)}
\format{
  A character matrix of size (number of individuals) times (number of markers in the genome).
}
\details{
 Note that there is only one column for each marker.
}
\examples{

data(geno)

}
