% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.sparse-data.R
\docType{data}
\name{geno.sparse-data}
\alias{geno.sparse-data}
\alias{geno.sparse}
\title{The Marker Genotypes of The Simulated Data (as sparse matrix)}
\format{
A sparse matrix of size (number of markers in the genome)
times (number of individuals).
}
\usage{
data(geno.sparse)
}
\description{
The marker genotypes of the simulated data
}
\details{
Note that there is only one column for each marker.
}
\examples{
data("geno.sparse", package="vGWAS")
}
\keyword{datasets}
