% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAF.R
\name{getMAF}
\alias{getMAF}
\title{Get minor-allele-frequency}
\usage{
getMAF(geno.matrix, geno.snp = "row", include.het = FALSE)
}
\arguments{
\item{geno.matrix}{a \code{matrix} or \code{data.frame} or
\code{sparseMatrix} with individuals as columns and markers as rows
(geno.snp = "row") or individuals as rows and markers as columns
(geno.snp = "col").}

\item{geno.snp}{if individuals at columns and markers at rows use "row" else
if individuals at rows and markers at columns use "col"}

\item{include.het}{specify if heterozygous calls should be split and added
equally to homozygous ref and alt counts (default = FALSE)}
}
\value{
a \code{vector} containing minor-allele-frequency values
}
\description{
Calculates minor-allele-frequency (MAF)
}
\examples{
data(geno)
maf <- getMAF(geno.matrix = geno, geno.snp = "col")
data(geno.num)
maf.num <- getMAF(geno.matrix = geno.num, geno.snp = "col")
data(geno.df)
maf.df <- getMAF(geno.matrix = geno.df, geno.snp = "row")
data(geno.sparse)
maf.sparse <- getMAF(geno.matrix = geno.sparse, geno.snp = "row")
}
\author{
Kristian Ullrich
}
