% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-data.R
\docType{data}
\name{map-data}
\alias{map-data}
\alias{map}
\title{Map Positions for The Markers of The Simulated Data}
\format{A numeric vector of chromosomal map positions of the 20K
simulated markers.}
\usage{
data(map)
}
\description{
Map positions for the markers of the simulated data
}
\examples{
data("map", package="vGWAS")
}
\keyword{datasets}
