\name{map}
\alias{map}
\docType{data}
\title{
Map Positions for The Markers of The Simulated Data
}
\description{
Map positions for the markers of the simulated data
}
\usage{data(chr)}
\format{
  A numeric vector of chromosomal map positions of the 20K simulated markers.
}
\examples{

data(map)

}
