% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-vGWAS.R
\name{package-vGWAS}
\alias{package-vGWAS}
\title{Variance Heterogeneity Genome-wide Association Study}
\description{
The package provides models and tests for variance
heterogeneity genome-wide association study (vGWAS).
"_PACKAGE"
}
\references{
Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O.
(2011): \bold{Inheritance beyond plain heritability:
variance-controlling genes in \emph{Arabidopsis thaliana}}.
\emph{PLoS Genetics}, \bold{8}, e1002839.\cr

Ronnegard, L., Shen, X. and Alam, M. (2011):
\bold{hglm: A Package for Fitting Hierarchical Generalized
Linear Models}. \emph{The R Journal}, \bold{2}(2), 20-28.\cr

Brown, M. B. and Forsythe, A.B. (1974).
\bold{Robust tests for equality of variances.} \emph{Journal
of the American Statistical Association}, \bold{69}, 364-367.\cr

Levene, H. (1960). \bold{Robust Tests for Equality
of Variances}, \emph{in Contributions to Probability and Statistics},
ed. I. Olkin, Palo Alto, CA: Stanford Univ. Press.\cr
}
\seealso{
R package \code{lawstat} for other types of nonparametric
variance tests and \code{onewaystats}.
}
\author{
Xia Shen
}
\keyword{GWAS}
