% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-data.R
\docType{data}
\name{pheno-data}
\alias{pheno-data}
\alias{pheno}
\title{Phenotypic Values for The Markers of The Simulated Data}
\format{A numeric vector of the phenotypic values of 93 simulated
individuals.}
\usage{
data(pheno)
}
\description{
Phenotypic values for the markers of the simulated data
}
\details{
Note that there is only one column for each marker.
}
\examples{
data("pheno", package="vGWAS")
hist(pheno, breaks = 30)
}
\keyword{datasets}
