\name{pheno}
\alias{pheno}
\docType{data}
\title{
Phenotypic Values for The Markers of The Simulated Data
}
\description{
Phenotypic values for the markers of the simulated data
}
\usage{data(pheno)}
\format{
  A numeric vector of the phenotypic values of 93 simulated individuals.
}
\examples{

data(pheno)
hist(pheno, breaks = 30)

}
\keyword{datasets}
