% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.vGWAS.R
\name{summary.vGWAS}
\alias{summary.vGWAS}
\title{Variance GWA Summary}
\usage{
\method{summary}{vGWAS}(object, nrMarkers = 10, ...)
}
\arguments{
\item{object}{a result object from \code{vGWAS} scan.
It can be any \code{list} or \code{data.frame} that contains
\code{chromosome}, \code{marker.map}, and \code{p.value},
with \code{class = 'vGWAS'}. See \code{\link{vGWAS}}.}

\item{nrMarkers}{a numeric value giving the number of top
markers to be summarized.}

\item{...}{not in use}
}
\value{
a summary for viewing vGWAS result.
}
\description{
The function summarized the variance GWA
result for the given scan object.
}
\examples{
\donttest{
# ----- load data ----- #
data(pheno)
data(geno)
data(chr)
data(map)
# ----- variance GWA scan ----- #
vgwa <- vGWAS(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr, pB = FALSE)
# ----- visualize the scan ----- #
plot(vgwa)
summary(vgwa)
# ----- calculate the variance explained by the strongest marker ----- #
vGWAS.variance(phenotype = pheno,
marker.genotype = geno[, vgwa[["p.value"]] == min(vgwa[["p.value"]])])
# ----- genomic control ----- #
vgwa2 <- vGWAS.gc(vgwa)
plot(vgwa2)
summary(vgwa2)
}
}
\references{
Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O.
(2011): \bold{Inheritance beyond plain heritability:
variance-controlling genes in \emph{Arabidopsis thaliana}}.
\emph{PLoS Genetics}, \bold{8}, e1002839.\cr

Ronnegard, L., Shen, X. and Alam, M. (2010):
\bold{hglm: A Package for Fitting Hierarchical Generalized
Linear Models}. \emph{The R Journal}, \bold{2}(2), 20-28.\cr
}
\seealso{
\code{\link{package-vGWAS}}, \code{\link{vGWAS}}
}
\author{
Xia Shen
}
