\name{vGWAS-package}
\alias{vGWAS-package}
\docType{package}
\title{
Variance Genome-wide Association
}
\description{
The package provides functions for genome-wide association using nonparametric variance test and some other further visualization and calculation. 
}
\details{
\tabular{ll}{
Package: \tab vGWAS\cr
Type: \tab Package\cr
Version: \tab 2010.10.12\cr
Date: \tab 2010-10-12\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab dglm\cr
}
}
\author{
Xia Shen

Maintainer: Xia Shen <xia.shen@lcb.uu.se>
}
\references{
Shen et al. (2010)
}
\keyword{GWAS}
\seealso{
package \code{lawstat} for other types of nonparametric variance tests.
}
