% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfmedian.test.R
\name{bfmedian.test}
\alias{bfmedian.test}
\alias{brown.forsythe.test}
\title{Brown-Forsythe's Test of Equality of Variances}
\usage{
bfmedian.test(y, group)
}
\arguments{
\item{formula}{a formula of the form lhs ~ rhs where lhs gives the sample
values and rhs the corresponding groups.}

\item{data}{a tibble or data frame containing the variables in formula.}

\item{alpha}{the level of significance to assess the statistical difference.
Default is set to alpha = 0.05.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds. Default us set to TRUE.}

\item{verbose}{a logical for printing output to R console.}
}
\value{
A list with class "owt" containing the following components:
\item{statistic}{the Brown-Forsythe test statistic.}
\item{parameter}{the parameter(s) of the approximate F distribution of the
test statistic.}
\item{p.value}{the p-value of the test.}
\item{alpha}{the level of significance to assess the statistical difference.}
\item{method}{the character string "Brown-Forsythe-Median Test".}
\item{data}{a data frame containing the variables in which NA values
(if exist) are removed.}
\item{formula}{a formula of the form lhs ~ rhs where lhs gives the sample
values and rhs the corresponding groups.}
}
\description{
The function performs the robust Brown-Forsythe
test using the group medians.
}
\details{
Levene (1960) proposed a test for homogeneity of variances
in \emph{k} groups which is based on the ANOVA statistic applied to
absolute deviations of observations from the corresponding group mean.
The robust Brown-Forsythe version of the Levene-type test substitutes
the group mean by the group median in the classical Levene statistic.
}
\note{
Modified from the \code{onewaytests} package and \code{vGWAS}.
}
\examples{
data(pheno)
data(geno)
df <- data.frame(phenotype = pheno, genotype = as.factor(geno[, 911]))
bfmedian.test(phenotype ~ genotype, data = df)
}
\references{
Brown, M. B. and Forsythe, A.B. (1974). \bold{Robust tests for
equality of variances.}
\emph{Journal of the American Statistical Association}, \bold{69},
364-367.\cr

Levene, H. (1960).
\bold{Robust Tests for Equality of Variances}, \emph{in Contributions
to Probability and Statistics}, ed. I. Olkin, Palo Alto, CA: Stanford Univ.
Press.\cr
}
\author{
Kristian Ullrich
}
\keyword{htest}
