% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGWAS.gc.R
\name{vGWAS.gc}
\alias{vGWAS.gc}
\title{Genomic Control for vGWAS}
\usage{
vGWAS.gc(object, plot = TRUE, proportion = 1, ...)
}
\arguments{
\item{object}{a result object from \code{vGWAS} scan.
It can be any \code{list} or \code{data.frame} that contains
\code{chromosome}, \code{marker.map}, and \code{p.value},
with \code{class = 'vGWAS'}. See \code{\link{vGWAS}}.}

\item{plot}{a logical value turning on/off the QQ plot
for genomic control.}

\item{proportion}{a numeric value between 0 and 1 giving
the proportion of obtained p-values to be used for genomic control.}

\item{...}{not in use}
}
\value{
\item{lambda}{estimated inflation ratio.}
\item{lambda.se}{standard error of the estimated inflation ratio.}
\item{gc.p.value}{p-values after genomic control.}
}
\description{
The function does genomic control for the
variance GWA result object.
}
\examples{
\donttest{
# ----- load data ----- #
data(pheno)
data(geno)
data(chr)
data(map)
# ----- variance GWA scan ----- #
vgwa <- vGWAS(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr, pb = FALSE)
# ----- visualize the scan ----- #
plot(vgwa)
summary(vgwa)
# ----- calculate the variance explained by the strongest marker ----- #
vGWAS.variance(phenotype = pheno,
marker.genotype = geno[,vgwa$p.value == min(vgwa$p.value)])
# ----- genomic control ----- #
vgwa2 <- vGWAS.gc(vgwa)
plot(vgwa2)
summary(vgwa2)
}
}
\references{
Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O.
(2011): \bold{Inheritance beyond plain heritability:
variance-controlling genes in \emph{Arabidopsis thaliana}}.
\emph{PLoS Genetics}, \bold{8}, e1002839.\cr
}
\seealso{
\code{\link{package-vGWAS}}, \code{\link{vGWAS}}
}
\author{
Xia Shen
}
