% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGWASparallel.R
\name{vGWASparallel}
\alias{vGWASparallel}
\title{Variance Genome-wide Association (parallel)}
\usage{
vGWASparallel(phenotype, geno.matrix, geno.snp = "row",
marker.map = NULL, chr.index = NULL, method = "bfmedian",
p.adjust.method = "none", include.het = FALSE, pB = TRUE, ncores = 1)
}
\arguments{
\item{phenotype}{a \code{numeric} or \code{logical} vector
of the phenotyic values.}

\item{geno.matrix}{a \code{matrix} or \code{data.frame} or
\code{sparseMatrix} with individuals as columns and markers as rows
(geno.snp = "row") or individuals as rows and markers as columns
(geno.snp = "col").}

\item{marker.map}{a \code{numeric} vector giving the marker map
positions for each chromosome.}

\item{chr.index}{a \code{numeric} vector giving the chromosome index
for each marker.}

\item{geno.snp}{if individuals at columns and markers at rows use "row" else
if individuals at rows and markers at columns use "col"}

\item{method}{the test method to use (default = bfmedian).
Default is set to the Brown-Forsythe's Test of Equality of Variances
using group medians.
There are 31 other tests available via the onewaytests package:
Alvandi's F test ("af"),
Alexander-Govern test ("ag"),
Alvandi's generalized p-value ("agp"),
One-way analysis of variance ("aov"),
Approximate F test ("ap"),
Adjusted Welch's heteroscedastic F test ("aw"),
B square test ("b2"),
Brown-Forsythe test ("bf"),
Box F test ("box"),
Cochran test ("cochran"),
Generalized tests equivalent to Parametric Bootstrap ("gtb"),
Generalized tests equivalent to Fiducial tests ("gtf"),
Variance homogeneity tests ("homog"),
James second order test ("james"),
Johansen F test ("johansen"),
Kruskal-Wallis test ("kw"),
Modified Brown-Forsythe test ("mbf"),
Mann-Whitney U test ("mw"),
Anderson-Darling normility test ("nor_ad"),
Cramer-vin Mises normility test ("nor_cvm"),
Kolmogorov-Smirnov normility test ("nor_ks"),
Pearson Chi-square normility test ("nor_pct"),
Shapiro-Wilk normility test ("nor_sw"),
Shapiro-Francia normility test ("nor_sf"),
Permutation F test ("pf"),
Scott-Smith test ("ss"),
Student's t-test ("st"),
Welch-Aspin test ("wa"),
Welch's heteroscedastic F test with trimmed means and Winsorized variances
("welch"),
Weerahandi's generalized F test ("wgf"),
Welch's t-test ("wt").}

\item{test.na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds. Default us set to TRUE.}

\item{p.adjust.method}{correction method (default = "none").
There are 8 p-value correction methods available via the p.adjust function:
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{include.het}{specify if heterozygous calls should be split and added
equally to homozygous ref and alt counts (default = FALSE)}

\item{pB}{show progress bar}

\item{ncores}{number of cores to parallelize (default = 1)}

\item{test.alpa}{the level of significance to assess the statistical
difference. Default is set to alpha = 0.05.}
}
\value{
a \code{data.frame} containing columns of \code{marker} names,
\code{chromosome} indices, \code{marker.map} positions,
test \code{statistic} values, and \code{p.value} for each position.
}
\description{
Variance Genome-wide association for using
nonparametric variance test and other
}
\examples{
# ----- load data ----- #
data(pheno)
data(geno)
data(chr)
data(map)
# ----- variance GWA scan ----- #
vgwa <- vGWASparallel(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr,
geno.snp = "col", pb = FALSE)
# ----- other test GWA scan ----- #
vgwa.mw <- vGWASparallel(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr,
geno.snp = "col", method = "mw", pb = FALSE)
# ----- multiple cores ----- #
vgwa.st <- vGWASparallel(phenotype = pheno, geno.matrix = geno,
marker.map = map, chr.index = chr,
geno.snp = "col", method = "st", ncores = 2, pb = FALSE)
}
\references{
Shen, X., Pettersson, M., Ronnegard, L. and Carlborg, O.
(2011): \bold{Inheritance beyond plain heritability:
variance-controlling genes in \emph{Arabidopsis thaliana}}.
\emph{PLoS Genetics}, \bold{8}, e1002839.\cr

Ronnegard, L., Shen, X. and Alam, M. (2010):
\bold{hglm: A Package for Fitting Hierarchical Generalized
Linear Models}. \emph{The R Journal}, \bold{2}(2), 20-28.\cr
}
\seealso{
\code{\link{package-vGWAS}} \code{onewaytests}
}
\author{
Xia Shen

Kristian Ullrich
}
